chore: squash old network config
