+++ src/components/records/E2DRecord.tsx
@@ -2,7 +2,7 @@ import { encodeAddress } from '@polkadot//util-crypto';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { RecordComponentProps } from '../../config';
-import { ApiKeys, EthereumConfig, Network, PolkadotConfig } from '../../model';
+import { EthereumConfig, Network, PolkadotChainConfig } from '../../model';
 import { E2DHistory as E2DRecordType, RedeemHistory, RingBurnHistory } from '../../model/darwinia';
 import { getLegalName, verticesToChainConfig } from '../../utils';
 import { Progresses, ProgressProps, State } from './Progress';
@@ -16,7 +16,7 @@ export function E2DRecord({
 }: RecordComponentProps<
   E2DRecordType & Partial<RingBurnHistory & RedeemHistory> & { isGenesis?: boolean },
   EthereumConfig,
-  PolkadotConfig<ApiKeys>
+  PolkadotChainConfig
 >) {
   const { chain, amount, currency, target, block_timestamp, is_relayed, tx, darwinia_tx, isGenesis } = record;
   const { t } = useTranslation();
