+++ src/config/bridge/ethereum-crabDVM.ts
@@ -1,6 +1,5 @@
-import { EthereumCrabDVMConfig } from '../../model';
-import { EVOLUTION_DOMAIN } from '../network';
-import { Bridge } from './bridge';
+import { Bridge, EthereumCrabDVMConfig } from '../../model';
+import { EVOLUTION_DOMAIN } from '../evolution';
 
 const ethereumCrabDVMConfig: EthereumCrabDVMConfig = {
   specVersion: 1180,
@@ -14,5 +13,6 @@ const ethereumCrabDVMConfig: EthereumCrabDVMConfig = {
 export const ethereumCrabDVM = new Bridge(
   { network: 'ethereum', mode: 'native' },
   { network: 'crab', mode: 'dvm' },
-  ethereumCrabDVMConfig
+  ethereumCrabDVMConfig,
+  { stable: false }
 );
