+++ src/config/evolution.ts
@@ -1,3 +1,8 @@
 export enum EvoApiPath {
   deposit = 'bank/gringotts',
 }
+
+export const EVOLUTION_DOMAIN = {
+  product: 'https://www.evolution.land',
+  dev: 'https://www.evolution.land.l2me.com',
+};
