+++ src/config/network/index.ts
@@ -1,7 +1,10 @@
-export * from './graph';
 export * from './crab';
 export * from './darwinia';
+export * from './descriptions';
 export * from './ethereum';
+export * from './graph';
+export * from './network-simple';
+export * from './networks';
 export * from './pangolin';
 export * from './pangoro';
 export * from './ropsten';
