+++ src/config/network/pangoro.ts
@@ -1,13 +1,6 @@
-import { EVOLUTION_DOMAIN } from '../network';
 import { PangoroConfig } from '../../model';
 
 export const pangoroConfig: PangoroConfig = {
-  api: {
-    dapp: 'https://api.darwinia.network.l2me.com',
-    evolution: EVOLUTION_DOMAIN.dev,
-    subGraph: 'https://pangolin-thegraph.darwinia.network/subgraphs/name/wormhole/Sub2SubMappingTokenFactory',
-    subql: 'https://api.subquery.network/sq/darwinia-network/wormhole-pangoro',
-  },
   facade: {
     logo: '/image/pangoro.png',
     logoMinor: '/image/pangoro.png',
