+++ src/config/network/tron.ts
@@ -1,9 +1,6 @@
 import { TronConfig } from '../../model';
 
 export const tronConfig: TronConfig = {
-  api: {
-    dapp: 'https://api.darwinia.network',
-  },
   facade: {
     logo: '/image/tron.png',
     logoMinor: '/image/tron.png',
