+++ src/model/index.ts
@@ -1,6 +1,7 @@
 export * from './account';
 export * from './bridge';
 export * from './common';
+export * from './connection';
 export * from './darwinia';
 export * from './erc20';
 export * from './evolution';
