+++ src/model/storage.ts
@@ -1,10 +1,10 @@
 import { THEME } from '../config';
 import { HashInfo } from '../utils';
-import { Network, NetworkConfig } from './network';
+import { Network, ChainConfig } from './network';
 
 export interface StorageInfo extends HashInfo {
   theme?: THEME;
   enableTestNetworks?: boolean;
-  config?: Partial<NetworkConfig>;
+  config?: Partial<ChainConfig>;
   custom?: Network[];
 }
